requirements.txt
Django==2.0.7
djangorestframework==3.8.2
django-celery-beat==1.1.1
channels==2.1.2
channels-redis==2.2.1
psycopg2-binary==2.7.5
requests==2.19.1
Celery==4.2.0
